ceph-volume: Refactor raw activate part
This move most of the logic from raw.activate.Activate
to objectstore.rawbluestore.RawBlueStore as part of the
major prepare/create/activate refactor.

Fixes: https://tracker.ceph.com/issues/61827
